+++ AeSys/EoGsViewTransform.h
@@ -3,10 +3,10 @@
 #include "EoGsAbstractView.h"
 
 class EoGsViewTransform : public EoGsAbstractView {
-  float m_UMin;
-  float m_VMin;
-  float m_UMax;
-  float m_VMax;
+  double m_UMin;
+  double m_VMin;
+  double m_UMax;
+  double m_VMax;
 
   EoGeTransformMatrix m_Matrix;
   EoGeTransformMatrix m_ProjectionMatrix;
@@ -23,7 +23,7 @@ class EoGsViewTransform : public EoGsAbstractView {
   EoGsViewTransform& operator=(const EoGsViewTransform& src);
 
  public:  // Methods
-  void AdjustWindow(const float aspectRatio);
+  void AdjustWindow(const double aspectRatio);
   void BuildTransformMatrix();
   EoGeTransformMatrix& GetMatrix();
   EoGeTransformMatrix& GetMatrixInverse();
@@ -32,20 +32,20 @@ class EoGsViewTransform : public EoGsAbstractView {
   void ZAxisRotation(double dSinAng, double dCosAng);
   void Scale(EoGeVector3d scale);
   /// <summary> Sets a window which is centered on the view target after adjusting for viewport aspect ratio</summary>
-  void SetCenteredWindow(const EoGsViewport& viewport, float uExtent, float vExtent);
+  void SetCenteredWindow(const EoGsViewport& viewport, double uExtent, double vExtent);
   void SetMatrix(EoGeTransformMatrix& tm);
-  void SetWindow(const float uMin, const float vMin, const float uMax, const float vMax);
+  void SetWindow(const double uMin, const double vMin, const double uMax, const double vMax);
   void TransformPoint(EoGePoint4d& point);
   void TransformPoints(EoGePoint4dArray& pointsArray);
   void TransformPoints(int numberOfPoints, EoGePoint4d* points);
   void TransformVector(EoGeVector3d& vector);
   void Translate(EoGeVector3d translate);
-  float UExtent() const;
-  float UMax() const;
-  float UMin() const;
-  float VExtent() const;
-  float VMax() const;
-  float VMin() const;
+  double UExtent() const;
+  double UMax() const;
+  double UMin() const;
+  double VExtent() const;
+  double VMax() const;
+  double VMin() const;
 };
 
 typedef CList<EoGsViewTransform> EoGsViewTransforms;
